Fix minor warnings/build issues
comm -> #ifndef NO_COMM

+++ enginee/src/main.cpp
@@ -175,7 +175,7 @@ int main(int argc, char** argv) {
         double total_time = frame_time - init_time;
         UNUSED(total_time);
         //frame_time.delta_s()
-        all_done = demo::run(1.f/60.f, total_frames/60.f, total_frames);
+        all_done = demo::run(1.f/60.f, static_cast<float>(total_frames)/60.f, total_frames);
         #ifndef NO_RENDER
         glfwSetWindowShouldClose(window, all_done);
         #endif
